+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Introduces a slight backward incompatibility. Need to import as: `from dse_do_ut
 - CplexDot in core01_optimization_engine: generic function class to use groupby aggregation and the mdl.dot() function.
 - PlotlyManager - self.ref_dm, self.ms_inputs, self.ms_outputs property declarations and documentation
 - PlotlyManager.plotly_kpi_compare_bar_charts and .get_multi_scenario_table for scenario compare
+- Core01DataManager and Core01OptimizationEngine: added support for parameter `mipGap`. Sets the `mdl.parameters.mip.tolerances.mipgap` if value > 0
 
 ## [0.5.6.0]- 2023-05-13
 ### Changed